refactor(auth): replace direct user model calls with useAuth utility for JWT token management
