+++ server/api/v1/auth/login-jwt.post.ts
@@ -24,9 +24,9 @@ export default defineEventHandler(async (event) => {
     throw createError({ statusCode: 401, message: 'Credenciais inválidas' })
   }
 
-  const token = await user.generateJWTToken(loginUser)
+  const token = await useAuth().generateJWTToken(loginUser)
 
-  const refreshToken = await user.generateJWTTokenRefresh(loginUser)
+  const refreshToken = await useAuth().generateJWTTokenRefresh(loginUser)
 
   const loginUserData = user.transformToLogin(loginUser)
