+++ server/utils/auth.ts
@@ -1,5 +1,8 @@
 import type { H3Event } from 'h3'
 import userModel from '#server/models/user'
+import type { User } from '#server/database/schemas/users'
+import type { UserLogin } from '#server/models/user'
+import * as jose from 'jose'
 
 const validateAuth = async (event: H3Event): Promise<{ email: string }> => {
   try {
@@ -16,7 +19,7 @@ const validateAuth = async (event: H3Event): Promise<{ email: string }> => {
 
   try {
     const token = bearerToken.split(' ')[1]
-    const decodedToken = await userModel.verifyJWTToken(token)
+    const decodedToken = await useAuth().verifyJWTToken(token)
     return decodedToken
   } catch (error) {
     throw createError({ statusCode: 401, message: 'Verificação JWT inválida', cause: error })
@@ -52,3 +55,62 @@ const requiredAuth = async (event: H3Event) => {
 }
 
 export { requiredAuth }
+
+export const useAuth = () => {
+  const secret = Buffer.from(process.env.NUXT_SESSION_PASSWORD!, 'utf-8')
+
+  const generateJWTToken = async (user: User): Promise<{ token: string; exp: number }> => {
+    const payload = userModel.transformToLogin(user, 'access')
+
+    const exp = new Date(Date.now() + 60 * 15 * 1000) // 15 minutes
+
+    const token = await new jose.SignJWT(payload as unknown as jose.JWTPayload)
+      .setProtectedHeader({ alg: 'HS256' })
+      .setExpirationTime(exp)
+      .sign(secret)
+
+    return { token, exp: exp.getTime() }
+  }
+
+  const generateJWTTokenRefresh = async (user: User): Promise<{ token: string; exp: number }> => {
+    const payload = userModel.transformToLogin(user, 'refresh')
+
+    const exp = new Date(Date.now() + 60 * 60 * 24 * 7 * 1000) // 7 days
+
+    const token = await new jose.SignJWT(payload as unknown as jose.JWTPayload)
+      .setProtectedHeader({ alg: 'HS256' })
+      .setExpirationTime(exp)
+      .sign(secret)
+
+    return { token, exp: exp.getTime() }
+  }
+
+  const verifyJWTToken = async (token: string): Promise<UserLogin> => {
+    const blacklist = ((await useStorage().getItem('jwt-blacklist')) as string[]) || []
+
+    if (blacklist.includes(token)) {
+      throw createError({ statusCode: 401, message: 'Token expirado ou inválido' })
+    }
+
+    try {
+      const { payload } = await jose.jwtVerify(token, secret)
+      return payload as unknown as UserLogin
+    } catch (error) {
+      console.error(error)
+      throw createError({ statusCode: 401, message: 'Token expirado ou inválido', cause: error })
+    }
+  }
+
+  const invalidateJWTToken = async (token: string): Promise<void> => {
+    const blacklist = ((await useStorage().getItem('jwt-blacklist')) as string[]) || []
+    blacklist.push(token)
+    await useStorage().setItem('jwt-blacklist', blacklist)
+  }
+
+  return {
+    generateJWTToken,
+    generateJWTTokenRefresh,
+    verifyJWTToken,
+    invalidateJWTToken
+  }
+}
